moved combatsystem to shared dll added a few more test skills fixed bugs in combat system added logging to console

+++ assailant/TimeKeeper.cs
@@ -12,7 +12,7 @@ namespace assailant
         static Stopwatch _Stopwatch = Stopwatch.StartNew();
         public static double hires_time_in_seconds()
         {
-            return _Stopwatch.ElapsedMilliseconds / 1000;
+            return _Stopwatch.ElapsedMilliseconds / (double)1000;
         }
     }
 }
